Project 4/requirements.txt
oracledb
uvicorn
fastapi
python-dotenv
pydantic
requests

